Added ability to view video
The camera doesn't appear to have RTSP although it has an open RTSP
port. The service there also speaks somewhat RTSP, but doesn't serve a
video stream (instead just hangs). The way the official app is getting
video is via a connection to the port 34567, where through some hoops a
multiplexed RTSP stream can be fetched. We do just that and pipe it into
mplayer, which after quite a while autodetects the H264-ES format and
then plays back the video.